+++ projects/tests/tests/test-remult-admin.spec.ts
@@ -372,11 +372,13 @@ describe('remult-admin', () => {
           "key": "accounts",
           "relations": [
             {
+              "caption": "Users",
               "compoundIdField": undefined,
               "entityKey": "users",
               "fields": {
                 "account": "id",
               },
+              "key": "users",
               "where": undefined,
             },
           ],
@@ -394,12 +396,16 @@ describe('remult-admin', () => {
     }
 
     const res = remultAdminHtml({
-      entities: [User],
-      remult: new Remult(),
+      rootPath: '/api',
+      head: '<title>Test Admin</title>',
     })
 
     expect(res).includes('html')
-    expect(res).includes('const optionsFromServer = ')
-    expect(res).not.includes('<!--PLACE_HERE-->')
+
+    expect(res).not.includes('<!--PLACE_HERE_HEAD-->')
+    expect(res).includes('<title>Test Admin</title>')
+
+    expect(res).not.includes('<!--PLACE_HERE_BODY-->')
+    expect(res).includes('window.optionsFromServer = ')
   })
 })